feat: Added BikeShare.py project.
This Python script is written for Project 2 (Term 1) of Udacity's Nanodegree and is used to explore data related to bike share systems for Chicago, New York City, and Washington. It imports data from csv files and compute descriptive statistics from the data. It also takes in users' raw input to create an interactive experience in the terminal to present these statistics.

Commit#1